# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vartalaap API",
    "description": "Voice bot platform for local Indian businesses",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/business/{business_id}": {
      "get": {
        "tags": [
          "Business",
          "business"
        ],
        "summary": "Get Business",
        "description": "Get business settings by ID.\n\nRequires X-Business-ID header matching the requested business_id.",
        "operationId": "get_business_api_business__business_id__get",
        "parameters": [
          {
            "name": "business_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Business Id"
            }
          },
          {
            "name": "x-business-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Business-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Business",
          "business"
        ],
        "summary": "Update Business",
        "description": "Update business settings.\n\nRequires X-Business-ID header matching the requested business_id.",
        "operationId": "update_business_api_business__business_id__patch",
        "parameters": [
          {
            "name": "business_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Business Id"
            }
          },
          {
            "name": "x-business-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Business-Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BusinessResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/BusinessType"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "status": {
            "$ref": "#/components/schemas/BusinessStatus"
          },
          "phone_numbers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Phone Numbers"
          },
          "operating_hours": {
            "additionalProperties": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OperatingHours"
                },
                {
                  "type": "string"
                }
              ]
            },
            "type": "object",
            "title": "Operating Hours"
          },
          "reservation_rules": {
            "$ref": "#/components/schemas/ReservationRules"
          },
          "greeting_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Greeting Text"
          },
          "menu_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Menu Summary"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "timezone",
          "status",
          "phone_numbers",
          "operating_hours",
          "reservation_rules",
          "greeting_text",
          "menu_summary"
        ],
        "title": "BusinessResponse",
        "description": "Business settings response."
      },
      "BusinessStatus": {
        "type": "string",
        "enum": [
          "active",
          "suspended",
          "onboarding"
        ],
        "title": "BusinessStatus",
        "description": "Business account status."
      },
      "BusinessType": {
        "type": "string",
        "enum": [
          "restaurant",
          "clinic",
          "salon",
          "other"
        ],
        "title": "BusinessType",
        "description": "Type of business."
      },
      "BusinessUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "phone_numbers": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Numbers"
          },
          "operating_hours": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/OperatingHours"
                    },
                    {
                      "type": "string"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operating Hours"
          },
          "reservation_rules": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReservationRules"
              },
              {
                "type": "null"
              }
            ]
          },
          "greeting_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Greeting Text"
          },
          "menu_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Menu Summary"
          }
        },
        "type": "object",
        "title": "BusinessUpdate",
        "description": "Business settings update request."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OperatingHours": {
        "properties": {
          "open": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open",
            "description": "Opening time (HH:MM) or null if closed"
          },
          "close": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close",
            "description": "Closing time (HH:MM) or null if closed"
          }
        },
        "type": "object",
        "title": "OperatingHours",
        "description": "Operating hours for a single day."
      },
...
      "ReservationRules": {
        "properties": {
          "min_party_size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Min Party Size",
            "default": 1
          },
          "max_party_size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Max Party Size",
            "default": 10
          },
          "max_phone_party_size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Max Phone Party Size",
            "description": "Max party size bookable by phone (larger \u2192 WhatsApp)",
            "default": 10
          },
          "total_seats": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Total Seats",
            "default": 40
          },
          "advance_days": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Advance Days",
            "description": "How many days ahead can book",
            "default": 30
          },
          "slot_duration_minutes": {
            "type": "integer",
            "minimum": 15.0,
            "title": "Slot Duration Minutes",
            "default": 90
          }
        },
        "type": "object",
        "title": "ReservationRules",
        "description": "Reservation rules for the business."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}